+++ core/src/main/java/com/graphhopper/util/EdgeIteratorState.java
@@ -22,7 +22,7 @@ import com.graphhopper.search.KVStorage;
 import com.graphhopper.storage.Graph;
 import com.graphhopper.storage.IntsRef;
 
-import java.util.List;
+import java.util.Map;
 
 /**
  * This interface represents an edge and is one possible state of an EdgeIterator.
@@ -146,8 +146,9 @@ public interface EdgeIteratorState {
 
     /**
      * @param list is a sorted collection of coordinates between the base node and the current adjacent node. Specify
-     *             the list without the adjacent and base node. This method can be called multiple times, but if the
-     *             distance changes, the setDistance method is not called automatically.
+     *             the list without the adjacent and base node. This method can be called multiple times, unless the
+     *             given point list is longer than the first time the method was called. Also keep in
+     *             mind that if the distance changes the setDistance method is not called automatically.
      */
     EdgeIteratorState setWayGeometry(PointList list);
 
@@ -233,14 +234,14 @@ public interface EdgeIteratorState {
      * But it might be slow and more inefficient on retrieval. Call this setKeyValues method only once per
      * EdgeIteratorState as it allocates new space everytime this method is called.
      */
-    EdgeIteratorState setKeyValues(List<KVStorage.KeyValue> map);
+    EdgeIteratorState setKeyValues(Map<String, KVStorage.KValue> map);
 
     /**
      * This method returns KeyValue pairs for both directions in contrast to {@link #getValue(String)}.
      *
-     * @see #setKeyValues(List)
+     * @see #setKeyValues(Map)
      */
-    List<KVStorage.KeyValue> getKeyValues();
+    Map<String, KVStorage.KValue> getKeyValues();
 
     /**
      * This method returns the *first* value for the specified key and only if stored for the direction of this